Added CreateSOAPEnvelopeHandlerTest. Refactored several tests.
