+++ modules/holodeckb2b-core/src/test/java/org/holodeckb2b/ebms3/handlers/outflow/CreateSOAPEnvelopeHandlerTest.java
@@ -1,20 +1,4 @@
-/*
- * Copyright (C) 2016 The Holodeck B2B Team, Sander Fieten
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.holodeckb2b.ebms3.handlers.inflow;
+package org.holodeckb2b.ebms3.handlers.outflow;
 
 import org.apache.axiom.om.OMElement;
 import org.apache.axiom.soap.SOAPEnvelope;
@@ -28,8 +12,6 @@ import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 import org.apache.log4j.spi.LoggingEvent;
 
-import org.holodeckb2b.common.messagemodel.AgreementReference;
-import org.holodeckb2b.common.messagemodel.TradingPartner;
 import org.holodeckb2b.common.messagemodel.UserMessage;
 import org.holodeckb2b.common.mmd.xml.MessageMetaData;
 import org.holodeckb2b.core.testhelpers.HolodeckB2BTestCore;
@@ -38,15 +20,10 @@ import org.holodeckb2b.ebms3.packaging.Messaging;
 import org.holodeckb2b.ebms3.packaging.SOAPEnv;
 import org.holodeckb2b.ebms3.packaging.UserMessageElement;
 import org.holodeckb2b.interfaces.core.HolodeckB2BCoreInterface;
-import org.holodeckb2b.interfaces.general.EbMSConstants;
 import org.holodeckb2b.interfaces.persistency.entities.IUserMessageEntity;
-import org.holodeckb2b.persistency.dao.UpdateManager;
-import org.holodeckb2b.pmode.helpers.Agreement;
 import org.holodeckb2b.pmode.helpers.Leg;
 import org.holodeckb2b.pmode.helpers.PMode;
-import org.holodeckb2b.pmode.helpers.PartnerConfig;
 
-import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -61,17 +38,17 @@ import java.io.File;
 import java.util.List;
 
 import static org.junit.Assert.*;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.atLeastOnce;
+import static org.mockito.Mockito.verify;
 
 /**
- * Created at 23:22 29.01.17
+ * Created at 23:42 29.01.17
  *
  * @author Timur Shakuov (t.shakuov at gmail.com)
  */
 @RunWith(MockitoJUnitRunner.class)
-public class FindPModesTest {
+public class CreateSOAPEnvelopeHandlerTest {
 
-    // Appender to control logging events
     @Mock
     private Appender mockAppender;
     @Captor
@@ -81,11 +58,11 @@ public class FindPModesTest {
 
     private static HolodeckB2BTestCore core;
 
-    private FindPModes handler;
+    private CreateSOAPEnvelopeHandler handler;
 
     @BeforeClass
-    public static void setUpClass() throws Exception {
-        baseDir = FindPModesTest.class.getClassLoader()
+    public static void setUpClass() {
+        baseDir = CreateSOAPEnvelopeHandlerTest.class.getClassLoader()
                 .getResource("handlers").getPath();
         core = new HolodeckB2BTestCore(baseDir);
         HolodeckB2BCoreInterface.setImplementation(core);
@@ -93,16 +70,10 @@ public class FindPModesTest {
 
     @Before
     public void setUp() throws Exception {
-        handler = new FindPModes();
-        // Adding appender to the FindPModes logger
+        handler = new CreateSOAPEnvelopeHandler();
         Logger logger = LogManager.getRootLogger();
         logger.addAppender(mockAppender);
         logger.setLevel(Level.DEBUG);
-    }
-
-    @After
-    public void tearDown() throws Exception {
-        LogManager.getRootLogger().removeAppender(mockAppender);
     }
 
     @Test
@@ -116,7 +87,7 @@ public class FindPModesTest {
         OMElement umElement = UserMessageElement.createElement(headerBlock, mmd);
 
         MessageContext mc = new MessageContext();
-        mc.setFLOW(MessageContext.IN_FLOW);
+        mc.setFLOW(MessageContext.OUT_FLOW);
 
         try {
             mc.setEnvelope(env);
@@ -125,57 +96,25 @@ public class FindPModesTest {
         }
 
         PMode pmode = new PMode();
-        pmode.setMep(EbMSConstants.ONE_WAY_MEP);
-        pmode.setMepBinding(EbMSConstants.ONE_WAY_PUSH);
-
-        PartnerConfig initiator = new PartnerConfig();
-        pmode.setInitiator(initiator);
-
-        PartnerConfig responder = new PartnerConfig();
-        pmode.setResponder(responder);
 
         Leg leg = new Leg();
         pmode.addLeg(leg);
 
-        UserMessage userMessage
-                = UserMessageElement.readElement(umElement);
+        UserMessage userMessage = UserMessageElement.readElement(umElement);
 
-        String msgId = userMessage.getMessageId();
-
-        TradingPartner sender = userMessage.getSender();
-        initiator.setRole(sender.getRole());
-        initiator.setPartyIds(sender.getPartyIds());
-
-        TradingPartner receiver = userMessage.getReceiver();
-        responder.setRole(receiver.getRole());
-        responder.setPartyIds(receiver.getPartyIds());
-
-        AgreementReference agreementReference =
-                userMessage.getCollaborationInfo().getAgreement();
-        String pmodeId = agreementReference.getPModeId();
-        String agreementRefName = agreementReference.getName();
-        String agreementRefType = agreementReference.getType();
-
-        pmode.setId(pmodeId);
-
-        Agreement agreement = new Agreement();
-        agreement.setName(agreementRefName);
-        agreement.setType(agreementRefType);
-        pmode.setAgreement(agreement);
-
-        // todo It seems strange that we need to set the PMode id value separately
-        // todo when it is contained within the agreement
-        // todo but if we don't set it the value returned by userMessage.getPModeId() is null now
+        String pmodeId =
+                userMessage.getCollaborationInfo().getAgreement().getPModeId();
+        // todo the pmodeId should be set to userMessage in the UserMessageElement.readElement() method
         userMessage.setPModeId(pmodeId);
+//        System.out.println("userMessage pmode id: " + userMessage.getPModeId());
+        pmode.setId(pmodeId);
 
         core.getPModeSet().add(pmode);
 
         // Setting input message property
-        UpdateManager updateManager = core.getUpdateManager();
-        System.out.println("um: " + updateManager.getClass());
         IUserMessageEntity userMessageEntity =
-                updateManager.storeIncomingMessageUnit(userMessage);
-        mc.setProperty(MessageContextProperties.IN_USER_MESSAGE,
+                core.getUpdateManager().storeIncomingMessageUnit(userMessage);
+        mc.setProperty(MessageContextProperties.OUT_USER_MESSAGE,
                 userMessageEntity);
 
         try {
@@ -185,12 +124,10 @@ public class FindPModesTest {
             fail(e.getMessage());
         }
 
-        // Checking log messages to make sure handler found pmode
         verify(mockAppender, atLeastOnce())
                 .doAppend(captorLoggingEvent.capture());
         List<LoggingEvent> events = captorLoggingEvent.getAllValues();
-        String expLogMsg = "Found P-Mode [" + pmode.getId()
-                + "] for User Message [" + msgId + "]";
+        String expLogMsg = "Added SOAP envelope to message context";
         boolean containsExpLogMsg = false;
         for(LoggingEvent e : events) {
             if(e.getLevel().equals(Level.DEBUG)) {
